refactor(scripts): group consistently with NPM script name
