+++ package.json
@@ -10,10 +10,10 @@
     "build": "NODE_ENV=production nx exec --verbose -- webpack --config scripts/webpack/webpack.prod.js",
     "build:nominify": "yarn run build -- --env noMinify=1",
     "build:stats": "NODE_ENV=production webpack --progress --config scripts/webpack/webpack.stats.js",
-    "codeowners-manifest": "node ./scripts/codeowners-manifest.js",
+    "codeowners-manifest": "node ./scripts/codeowners-manifest/index.js",
     "codeowners-manifest:clean": "rm -rf codeowners-manifest && mkdir -p codeowners-manifest",
-    "codeowners-manifest:metadata": "node ./scripts/generate-codeowners-metadata.js",
-    "codeowners-manifest:raw": "node ./scripts/generate-codeowners-raw-audit.js",
+    "codeowners-manifest:metadata": "node ./scripts/codeowners-manifest/metadata.js",
+    "codeowners-manifest:raw": "node ./scripts/codeowners-manifest/raw.js",
     "dev": "NODE_ENV=dev nx exec -- webpack --config scripts/webpack/webpack.dev.js",
     "e2e": "./e2e/start-and-run-suite",
     "e2e:old-arch": "./e2e/start-and-run-suite old-arch",
